[FIRRTL] InnerSymbolTable: move to header
